Fix code block paste button position
When inside a block quote the paste icon would appear slighty outside
the code block.

The button should be in the same position with the exception that it
is one or two pixels more inside the code block. In other words it
aligns better with the border around the code block